Raise pod version to 1.1.1.

+++ Pod/Supervision/TBActorSupervisor.m
@@ -129,7 +129,7 @@ static NSString * const TBAKActorSupervisorQueue = @"com.jkrumow.ActorKit.TBActo
         operation.invocation.target = newActor;
         newActor.loadCount = @(newActor.loadCount.unsignedIntegerValue + 1);
     }
-    NSLog(@"Transferred actor mailbox with %i messages to actor <%p>", newActor.loadCount.unsignedIntegerValue, newActor);
+    NSLog(@"Transferred actor mailbox with %lu messages to actor <%p>", (unsigned long)newActor.loadCount.unsignedIntegerValue, newActor);
 }
 
 - (void)_transferMailboxesFromPool:(TBActorPool *)pool toPool:(TBActorPool *)newPool
